refactor(workbench): 代码架构重构 + 性能优化
性能优化:
- WorkbenchPage/MessageList/ChatInput 等组件改为逐字段 store 选择器，消除全量订阅导致的不必要重渲染
- SessionSlice 独立为 session-slice.ts，每个 slice 提供自己的 reset 方法，解耦 reset 函数
- 统一 messages 操作为 appendMessages/replaceMessages/clearMessages，消除 3 个 slice 并发写入的耦合

代码架构:
- MessageBubble 814 行拆分为 10 个独立组件文件（MarkdownContent, LegalText, UserMessageContent, BatchItemContent, InlineToolCalls, StreamingBubble, MessageActions, BatchDownloadButton, AssistantMeta）
- WorkbenchPage 内联的 SessionItem/EditableTitle 外移为独立组件
- API 层封装为 workbenchApi 统一对象导出
- index.ts 补全导出（组件/hooks/types/API）
