+++ frontend/src/features/workbench/index.ts
@@ -1,5 +1,41 @@
 /** 工作台模块入口 */
 
+// 页面
 export { WorkbenchPage } from './WorkbenchPage'
+
+// Store
 export { useWorkbenchStore } from './stores/workbench-store'
-export type { WorkbenchSession, WorkbenchMessage, LLMModel, SSEEvent } from './types'
+
+// API
+export { workbenchApi } from './api'
+
+// 组件
+export { ChatInput } from './components/ChatInput'
+export { ModelSelector } from './components/ModelSelector'
+export { ContextUsageBar } from './components/ContextUsageBar'
+
+// Hooks
+export { useContextUsage } from './hooks/use-context-usage'
+
+// Types
+export type {
+  WorkbenchSession,
+  WorkbenchMessage,
+  LLMModel,
+  ModelsResponse,
+  AgentType,
+  AgentInfo,
+  ApprovalState,
+  SSEEvent,
+  StreamingMessage,
+  ToolCallState,
+  BatchJob,
+  BatchJobItem,
+  BatchJobStatus,
+  BatchProgress,
+  FailedItemDetail,
+  Attachment,
+} from './types'
+
+// Constants
+export { AGENT_OPTIONS } from './types'
